+++ lib/src/ui/widgets/paginator_content/number_content.dart
@@ -3,7 +3,7 @@ import 'dart:math';
 import 'package:auto_size_text/auto_size_text.dart';
 import 'package:flutter/material.dart';
 import 'package:number_paginator/src/ui/widgets/inherited_number_paginator.dart';
-import 'package:number_paginator/src/ui/widgets/paginator_button.dart';
+import 'package:number_paginator/src/ui/widgets/buttons/paginator_button.dart';
 
 class NumberContent extends StatelessWidget {
   final int currentPage;
